pyproject.toml
[tool.poetry]
name = "mirea-schedule-telegram-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
aiogram = "^3.4.1"
loguru = "^0.7.2"
httpx = "^0.27.0"
emoji = "^2.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.test.dependencies]
pytest = "^7.3.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"